commit from laptop

+++ The game/Assets/scripts/healthControler.cs
@@ -7,8 +7,8 @@ public class healthControler : MonoBehaviour
 {
     Transform mainCam;
     [SerializeField] private healthBar hBar;
-	private Animator animator;
-	private Animator camAnim;
+    private Animator animator;
+    private Animator camAnim;
     private playerController controller;
     // public GameObject blood;
     // Sprite bloodV;
@@ -23,34 +23,41 @@ public class healthControler : MonoBehaviour
     public float health;
 
 
-    private void Start() {
+    private void Start()
+    {
         health = startHealth;
         mainCam = Camera.main.transform;
         // making health bar green if not enemy
-        if(gameObject.GetComponent<Buggle>()!=null){
-            if(!gameObject.GetComponent<Buggle>().enemy){
+        if (gameObject.GetComponent<Buggle>() != null)
+        {
+            if (!gameObject.GetComponent<Buggle>().enemy)
+            {
                 gameObject.transform.Find("healthBar").Find("bar").Find("barSprite").GetComponent<SpriteRenderer>().color = Color.green;
             }
         }
 
     }
-    private void Update() {
+    private void Update()
+    {
         controlBloodyDeath();
         faceCamera();
     }
-    void faceCamera(){
+    void faceCamera()
+    {
         var thisHealthBar = gameObject.transform.Find("healthBar");
-        if(thisHealthBar)
+        if (thisHealthBar)
             thisHealthBar.transform.LookAt(thisHealthBar.transform.position + mainCam.rotation * Vector3.back, mainCam.rotation * Vector3.up);
     }
-    public void takeDmg(float bulletDmg){
+    public void takeDmg(float bulletDmg)
+    {
         animator = gameObject.GetComponent<Animator>();
         camAnim = Camera.main.GetComponent<Animator>();
         animator.SetBool("hit", true);
 
-        if(gameObject.tag=="Player"){
+        if (gameObject.tag == "Player")
+        {
             camAnim.SetBool("shake", true);
-            if(alive)
+            if (alive)
                 bloodOpacity = .4f;
             StartCoroutine(stopHurt(1));
         }
@@ -60,71 +67,86 @@ public class healthControler : MonoBehaviour
         var playerDefence = gameObject.GetComponent<playerController>();
         var defence = 0;
         // this will make sure that we get the defence of the player or the buggle
-        if(buggleDefence != null){
+        if (buggleDefence != null)
+        {
             defence = buggleDefence.defence;
-        } else if(playerDefence != null) {
+        }
+        else if (playerDefence != null)
+        {
             defence = playerDefence.defence;
         }
         var effectiveDamage = bulletDmg - defence;
 
         // Debug.Log(effectiveDamage);
 
-        if(effectiveDamage<0)
+        if (effectiveDamage < 0)
             effectiveDamage = 0;
-        if(alive)
+        if (alive)
             health = health - effectiveDamage;
-        if(health <= 0){
+        if (health <= 0)
+        {
             health = 0;
-            
-            if(gameObject.tag=="Player")
+
+            if (gameObject.tag == "Player")
                 death();
             else
                 Destroy(gameObject);
-        } else {
+        }
+        else
+        {
             animator.SetBool("hit", false);
         }
-        
+
         // Debug.Log(health);
         updateBar();
     }
 
 
 
-    public void updateBar(){
-        if(hBar != null)
+    public void updateBar()
+    {
+        if (hBar != null)
             hBar.setSize(health / startHealth);
 
     }
-    void death(){
+    void death()
+    {
         alive = false;
         // animator = gameObject.GetComponent<Animator>();
         animator.SetBool("dead", true);
         animator.Play("death", 0);
-        controller = gameObject.GetComponent<playerController> ();
-        controller.drop();
+        controller = gameObject.GetComponent<playerController>();
+        if (controller.isHolding)
+        {
+            controller.drop();
+        }
         Destroy(GameObject.Find("healthBar"));
         controller.enabled = false;
     }
 
 
     IEnumerator stopHurt(float time)
- {
-     yield return new WaitForSeconds(time/10);
- 
-    camAnim.SetBool("shake", false);
- }
+    {
+        yield return new WaitForSeconds(time / 10);
+        camAnim.SetBool("shake", false);
+    }
 
-    void controlBloodyDeath(){
-        if(blood != null){
-            if(alive){
+    void controlBloodyDeath()
+    {
+        if (blood != null)
+        {
+            if (alive)
+            {
                 bloodOpacity -= .05f;
                 deathOpacity = 0;
                 blood.color = new Color(1f, 1f, 1f, bloodOpacity);
-            } else{
-                if(bloodOpacity<.5f)
+            }
+            else
+            {
+                if (bloodOpacity < .5f)
                     bloodOpacity += .002f;
-                    
-                if(deathOpacity<.6f)
+
+                if (deathOpacity < .6f)
                     deathOpacity += .002f;
 
                 blood.color = new Color(0f, 0f, 0f, bloodOpacity);
